+++ CHANGELOG.md
@@ -26,6 +26,7 @@ All notable changes to this project will be documented in this file.
 - [ConsoleUI] Recommendations of full changeset in ConsoleUI (#4416 by: HebaruSan)
 - [GUI] Fix GUI exceptions for module not in available modules list (#4431 by: HebaruSan)
 - [Core] Don't clone with symlinks on Linux (#4441 by: HebaruSan)
+- [Core] Use temp dir in instance for mod file transactions (#4462 by: HebaruSan)
 
 ### Internal